requirements.txt
sqlalchemy>=2.0.0
typing-extensions>=4.0.0 
python-dotenv>=1.0.0
requests>=2.31.0
pandas>=2.1.4